lib: Split out cfs writer code from lcfs-writer.c
This moves all the cfs specific code to lcfs-writer-cfs.c and
makes the required shared functions accessible in lcfs-internal.h.
