+++ libcomposefs/Makefile-lib.am
@@ -9,6 +9,7 @@ libcomposefs_la_SOURCES = \
                         $(COMPOSEFSDIR)/lcfs-fsverity.c \
                         $(COMPOSEFSDIR)/lcfs-fsvertiy.h \
                         $(COMPOSEFSDIR)/lcfs-writer-erofs.c \
+                        $(COMPOSEFSDIR)/lcfs-writer-cfs.c \
                         $(COMPOSEFSDIR)/lcfs-writer.c \
                         $(COMPOSEFSDIR)/lcfs-writer.h \
                         $(COMPOSEFSDIR)/lcfs-mount.c \
